Amadeus_Projet_1/requirements.txt
customtkinter
openai
elevenlabs
speechrecognition
pyaudio
sounddevice
numpy
python-dotenv
pyvts
openai
elevenlabs
colorama
customtkinter
speechrecognition
asyncio
os